SqlDust supports path-notation
When fields are passed to SqlDust with brackets, then they will be seen
as part of a path. This means that these fields will not be scanned as
reserved keyword, or as a number.
These fields will be prepended with a table-alias. and will result in
JOINS when necessary.

Example:
SELECT [key], [regexp] as regexp FROM users